'or even decrease how many \glspl{fatigue} they can regenerate while resting by sealing up their mouth' - could be clarified that this is due to target's inability to eat; this might not be obvious to everyone immediately.
'a gnome with a \gls{weight} of 4 would have a \gls{tn} of 6 to move through a portal, while human with Strength +1 would be at \tn[12]' - example uses weight for a Gnome, but Strength for a Human. It should use just weight for both, for clarity of the example. Otherwise, someone reading this rule would need to reference another rule ( how weight is partially based on Strength ) just to understand the example.