make human characters more common
When random characters were human on the roll of 6-8, this made humans
common, but not that common.  Basically every group would be around half
non-human.

Too many strange characters threatens to give the setting a high-fantasy
feel.  It also works less well with the idea of the Night Guard being
the default, since it's a human institution.