tighten guild and culture ties in stories
Players have a tiny chance of rolling up a gnome as their first
character, which makes gnomes special, but it puts up a harsh barrier if
you want to stick with a rules-as-written interpretation.

This small rewrite lays out a new path: spend a Story Point to introduce
some Gnomish culture in your background, then introduce a Gnomish ally
with another Story Point.  A third Story Point adds that ally as a
playable character -  PC in the Character Pool.

Introducing other guilds works similarly - one point to introduce the
ally, and another to add them to the character pool and have them join
as a second PC for a session.

Hopefully the time-buffers provide space for people to read a few lines
on the other guilds or cultures in their local area.